Sur-constrain the walk of romeo.
Add a task to limit the orientation of the waist.
Add a task to contol the orientation of the head.
Add a task to control the position of the arms.

Also the waist does not follow the legs naturally.
Due to the com constraint, the robot tends to lean
its head backward, and move its hands backward.