Correct the prototype of the method getClassName.
Use the dedicated macro to this purpose.
The error came from the fact that the method was not declared
as const. The consequences were that the following python orders
were possible:

mytask = TaskInequality ('mytask')
mytask = Task ('mytask')

The second line should crash but doesn't, because the method getClassName
called is the one of the mother class.